+++ NEWS.md
@@ -44,6 +44,9 @@
 + IMPROVEMENT: Broker configuration now, like hook configuration, allows the
   usage of the `default` keyword, indicating the starting value for a
   configuration property if not overridden upon creation.
++ IMPROVEMENT: The Windows build-winpe step now allows the addition of extra
+  drivers to the generated .wim file. The drivers (.inf extension) need to be
+  added to the `extra-drivers` folder inside the build-winpe directory.
 + IMPROVEMENT: Line endings for SetupComplete.cmd.erb are now in Windows format,
   causing the rendered SetupComplete.cmd file to be legible on Windows systems.
 + IMPROVEMENT: The RAZOR_HTTP_PORT environment variable will now be used to